move retry logic to Resume(), move Namespace() to controller
