+++ metricproviders/datadog/datadog.go
@@ -3,12 +3,10 @@ package datadog
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"strings"
 	"time"
 
 	"github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1"
+	"github.com/argoproj/argo-rollouts/utils/controller"
 	"github.com/argoproj/argo-rollouts/utils/evaluate"
 	metricutil "github.com/argoproj/argo-rollouts/utils/metric"
 	templateutil "github.com/argoproj/argo-rollouts/utils/template"
@@ -56,19 +54,6 @@ func (p *Provider) Run(run *v1alpha1.AnalysisRun, metric v1alpha1.Metric) v1alph
 		return metricutil.MarkMeasurementError(newMeasurement, err)
 	}
 
-	// If datadog returns an empty []Series, retry with a longer window
-	var retries int
-	for retries = 0; retries < 3 && len(response) < 1; retries++ {
-		response, err = p.client.QueryMetrics(from-60, to, query)
-		if err != nil {
-			return metricutil.MarkMeasurementError(newMeasurement, err)
-		}
-
-		if len(response) < 1 {
-			time.Sleep(1 * time.Second)
-		}
-	}
-
 	if len(response) > 0 {
 		newValue, newStatus, err := p.processResponse(metric, response[0])
 		if err != nil {
@@ -87,9 +72,17 @@ func (p *Provider) Run(run *v1alpha1.AnalysisRun, metric v1alpha1.Metric) v1alph
 	return newMeasurement
 }
 
-// Resume should not be used with the datadog provider since all the work should occur in the Run method
+// Resume executes Run if the analysis phase is inconclusive or until 3 retries have been executed
 func (p *Provider) Resume(run *v1alpha1.AnalysisRun, metric v1alpha1.Metric, measurement v1alpha1.Measurement) v1alpha1.Measurement {
-	p.logCtx.Warn("Datadog provider should not execute the Resume method")
+	measurement = p.Run(run, metric)
+	var retries int
+	for retries = 0; retries < 3 && measurement.Phase == v1alpha1.AnalysisPhaseInconclusive; retries++ {
+		measurement = p.Run(run, metric)
+		if measurement.Value != "" {
+			time.Sleep(1 * time.Second)
+		}
+		return measurement
+	}
 	return measurement
 }
 
@@ -104,20 +97,27 @@ func (p *Provider) GarbageCollect(run *v1alpha1.AnalysisRun, metric v1alpha1.Met
 	return nil
 }
 
+// flattenResponse removes the unix timestamp from a []DataPoint and flattens all values into a []float64
+func flattenResponse(dp []dd.DataPoint) []float64 {
+	flat := make([]float64, len(dp))
+	for _, v := range dp {
+		flat = append(flat, *v[1])
+	}
+	return flat
+}
+
 func (p *Provider) processResponse(metric v1alpha1.Metric, response dd.Series) (string, v1alpha1.AnalysisPhase, error) {
 	length := len(response.Points)
 	switch {
 	case length < 1:
 		return "", v1alpha1.AnalysisPhaseInconclusive, nil
-	// If we get 1 or more points, select the first point
 	case length >= 1:
-		result := *response.Points[0][1]
+		result := flattenResponse(response.Points)
 		valueStr := fmt.Sprintf("%f", result)
 		newStatus := evaluate.EvaluateResult(result, metric, p.logCtx)
 		return valueStr, newStatus, nil
-	// TODO (cabrinha) add other response types
 	default:
-		return "", v1alpha1.AnalysisPhaseError, fmt.Errorf("Datadog metric type not supported")
+		return "", v1alpha1.AnalysisPhaseError, fmt.Errorf("failed to process metrics")
 	}
 }
 
@@ -131,7 +131,7 @@ func NewDatadogProvider(client dd.Client, logCtx log.Entry) *Provider {
 
 // NewDatadogAPI generates a datadog API from the metric configuration
 func NewDatadogAPI(metric v1alpha1.Metric, kubeclientset kubernetes.Interface) (*dd.Client, error) {
-	ns := Namespace()
+	ns := controller.Namespace()
 	secret, err := kubeclientset.CoreV1().Secrets(ns).Get(DatadogSecretName, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
@@ -153,20 +153,4 @@ func NewDatadogAPI(metric v1alpha1.Metric, kubeclientset kubernetes.Interface) (
 		return client, nil
 	}
 	return nil, errors.New("failed to make client: no datadog API or App keys found")
-}
-
-// Namespace will return $POD_NAMESPACE if set, otherwise "argo-rollouts"
-func Namespace() string {
-	// This way assumes you've set the POD_NAMESPACE environment variable using the downward API.
-	// This check has to be done first for backwards compatibility with the way InClusterConfig was originally set up
-	if ns, ok := os.LookupEnv("POD_NAMESPACE"); ok {
-		return ns
-	}
-	// Fall back to the namespace associated with the service account token, if available
-	if data, err := ioutil.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace"); err == nil {
-		if ns := strings.TrimSpace(string(data)); len(ns) > 0 {
-			return ns
-		}
-	}
-	return "argo-rollouts"
 }
