+++ utils/controller/controller.go
@@ -2,7 +2,10 @@ package controller
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
 	"runtime/debug"
+	"strings"
 	"time"
 
 	log "github.com/sirupsen/logrus"
@@ -250,4 +253,20 @@ func InstanceIDRequirement(instanceID string) labels.Requirement {
 		panic(err)
 	}
 	return *instanceIDReq
+}
+
+// Namespace will return $POD_NAMESPACE if set, otherwise "argo-rollouts"
+func Namespace() string {
+	// This way assumes you've set the POD_NAMESPACE environment variable using the downward API.
+	// This check has to be done first for backwards compatibility with the way InClusterConfig was originally set up
+	if ns, ok := os.LookupEnv("POD_NAMESPACE"); ok {
+		return ns
+	}
+	// Fall back to the namespace associated with the service account token, if available
+	if data, err := ioutil.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace"); err == nil {
+		if ns := strings.TrimSpace(string(data)); len(ns) > 0 {
+			return ns
+		}
+	}
+	return "argo-rollouts"
 }
